03-webapp/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1

